+++ modules-services2/service-templates-core/src/main/java/org/openprovenance/prov/service/core/readers/TemplatesVizConfig.java
@@ -1,8 +1,11 @@
 package org.openprovenance.prov.service.core.readers;
 
+import java.util.Map;
+
 public class TemplatesVizConfig {
     public Integer id;
     public String template;
     public String property;
     public String style;
+    public Map<String, String> parameters;
 }